requirements.txt
Django==1.10.5
django-bootstrap3==8.1.0
django-braces
django-autoslug==1.9.3
django-debug-toolbar==1.6
djangorestframework==3.5.4